Organize AArch64 neon tests
